Rename to namespace package cegal.welltools
